chore: simplify payout tokens handling
